+++ source/standalone_examples/testing/isaacsim.core.api/rigid_prim_view.py
@@ -40,8 +40,6 @@ my_cloner = Cloner()
 my_world = World(stage_units_in_meters=1.0, backend="torch")
 my_world.scene.add_default_ground_plane()
 
-asset_path = assets_root_path + "/Isaac/Robots/FrankaRobotics/FrankaPanda/franka_alt_fingers.usd"
-
 cube = DynamicCuboid(prim_path="/World/cube_0")
 prim_paths = my_cloner.generate_paths("/World/cube", 3)
 my_cloner.clone(cube.prim_path, prim_paths)